# api-catalogo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio de Catálogo de Homelab Collector Platform.",
        "title": "Homelab Collector Platform - Catalogo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9094",
    "basePath": "/",
    "paths": {
        "/api/v1/catalogo/especies": {
            "get": {
                "description": "Obtiene la lista de todas las especies de árbol registradas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Especies de Árbol"
                ],
                "summary": "Listar especies de árbol",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.EspecieArbol"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea una nueva especie de árbol en el catálogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Especies de Árbol"
                ],
                "summary": "Crear especie de árbol",
                "parameters": [
                    {
                        "description": "Datos para crear la especie",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateEspecieArbolRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EspecieArbol"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/catalogo/especies/{id}": {
            "get": {
                "description": "Obtiene los detalles de una especie de árbol por su UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Especies de Árbol"
                ],
                "summary": "Obtener especie de árbol por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la especie (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EspecieArbol"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza los datos de una especie de árbol existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Especies de Árbol"
                ],
                "summary": "Actualizar especie de árbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la especie (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEspecieArbolRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EspecieArbol"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina una especie de árbol del catálogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Especies de Árbol"
                ],
                "summary": "Eliminar especie de árbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la especie (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.CreateEspecieArbolRequest": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "descripcion": {
                    "type": "string"
                },
                "familia": {
                    "type": "string"
                },
                "nombre_cientifico": {
                    "type": "string"
                },
                "nombre_comun": {
                    "type": "string"
                }
            }
        },
...
        "model.EspecieArbol": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "descripcion": {
                    "type": "string"
                },
                "familia": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nombre_cientifico": {
                    "type": "string"
                },
                "nombre_comun": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateEspecieArbolRequest": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "descripcion": {
                    "type": "string"
                },
                "familia": {
                    "type": "string"
                },
                "nombre_cientifico": {
                    "type": "string"
                },
                "nombre_comun": {
                    "type": "string"
                }
            }
        },
...
        "response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}